pyproject.toml
[project]
name = "eval-sufficiency"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13,<1.14",
    "datasets==4.8.4",
    "ipykernel>=7.2.0",
    "numpy==2.2.6",
    "pandas==2.3.3",
    "pillow==11.1.0",
    "transformers==4.57.6",
    "torch==2.9.1",
    "torchaudio==2.9.1",
    "torchvision==0.24.1",
    "nvitop>=1.6.2",
    "matplotlib>=3.10.8",
]

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
torchaudio = { index = "pytorch-cu130" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
